surrealdb_protocol::fb

Module v1

Structs

AnyKind

Represents the ‘any’ type, which matches any value.

AnyKindArgs

AnyKindBuilder

Array

A collection of values.

ArrayArgs

ArrayBuilder

ArrayKind

Represents the ‘array’ type.

ArrayKindArgs

ArrayKindBuilder

BoolKind

Represents the ‘bool’ type.

BoolKindArgs

BoolKindBuilder

BoolValue

Boolean value.

BoolValueArgs

BoolValueBuilder

Bytes

Bytes value.

BytesArgs

BytesBuilder

BytesKind

Represents the ‘bytes’ type.

BytesKindArgs

BytesKindBuilder

DatetimeKind

Represents the ‘datetime’ type.

DatetimeKindArgs

DatetimeKindBuilder

Decimal

Decimal value represented as a string.

DecimalArgs

DecimalBuilder

DecimalKind

Represents the ‘decimal’ type.

DecimalKindArgs

DecimalKindBuilder

Duration

Duration value.

DurationArgs

DurationBuilder

DurationKind

Represents the ‘duration’ type.

DurationKindArgs

DurationKindBuilder

EitherKind

Represents the ‘either’ type (a union of multiple types).

EitherKindArgs

EitherKindBuilder

File

A file in a bucket.

FileArgs

FileBuilder

FileKind

Represents the ‘file’ type, optionally constrained to specific buckets.

FileKindArgs

FileKindBuilder

Float64Value

64-bit floating point value.

Float64ValueArgs

Float64ValueBuilder

FloatKind

Represents the ‘float’ type.

FloatKindArgs

FloatKindBuilder

FunctionKind

Represents the ‘function’ type with argument types and return type.

FunctionKindArgs

FunctionKindBuilder

Geometry

A Geometry object.

GeometryArgs

GeometryBuilder

GeometryCollection

A collection of Geometry objects.

GeometryCollectionArgs

GeometryCollectionBuilder

GeometryKind

Represents the ‘geometry’ type, optionally constrained to specific geometry types.

GeometryKindArgs

GeometryKindBuilder

GeometryKindType

Enum of geometry types for type constraints.

GeometryType

A union of all geometry types.

GeometryTypeUnionTableOffset

Int64Value

64-bit integer value.

Int64ValueArgs

Int64ValueBuilder

IntKind

Represents the ‘int’ type.

IntKindArgs

IntKindBuilder

KeyValue

A key-value pair.

KeyValueArgs

KeyValueBuilder

Kind

A type kind container that represents a type in SurrealDB’s type system.

KindArgs

KindBuilder

KindType

A union of all possible type kinds in SurrealDB’s type system.

KindTypeUnionTableOffset

LineString

An ordered collection of Points, representing a path between locations. To be valid, a LineString must be empty, or have two or more coords.

LineStringArgs

LineStringBuilder

LiteralArray

A literal array type with specific kinds for each element.

LiteralArrayArgs

LiteralArrayBuilder

LiteralDiscriminatedObject

A discriminated object type (tagged union) with variants distinguished by a discriminant field.

LiteralDiscriminatedObjectArgs

LiteralDiscriminatedObjectBuilder

LiteralKind

Represents a literal type (specific constant values or structured types).

LiteralKindArgs

LiteralKindBuilder

LiteralObject

A literal object type with specific fields and their types.

LiteralObjectArgs

LiteralObjectBuilder

LiteralType

A union of all possible literal type values.

LiteralTypeUnionTableOffset

MultiLineString

A collection of LineStrings.

MultiLineStringArgs

MultiLineStringBuilder

MultiPoint

A collection of Points.

MultiPointArgs

MultiPointBuilder

MultiPolygon

A collection of Polygons.

MultiPolygonArgs

MultiPolygonBuilder

NullKind

Represents the ‘null’ type.

NullKindArgs

NullKindBuilder

NullValue

Null value.

NullValueArgs

NullValueBuilder

NumberKind

Represents the ‘number’ type (includes int, float, and decimal).

NumberKindArgs

NumberKindBuilder

Object

A collection of key-value pairs.

ObjectArgs

ObjectBuilder

ObjectField

A field in a literal object type definition.

ObjectFieldArgs

ObjectFieldBuilder

ObjectKind

Represents the ‘object’ type.

ObjectKindArgs

ObjectKindBuilder

OptionKind

Represents the ‘option’ type (nullable/optional values).

OptionKindArgs

OptionKindBuilder

Point

A single point in 2D space.

PointArgs

PointBuilder

Polygon

A bounded two-dimensional area.

PolygonArgs

PolygonBuilder

Range

A range with start and end bounds.

RangeArgs

RangeBuilder

RangeKind

Represents the ‘range’ type.

RangeKindArgs

RangeKindBuilder

RecordId

An identifier for a record in a table.

RecordIdArgs

RecordIdBuilder

RecordIdKey

A record ID key that can be an integer, string, UUID, array, or range.

RecordIdKeyArgs

RecordIdKeyBound

A bound for a record ID key range, which can be inclusive, exclusive, or unbounded.

RecordIdKeyBoundUnionTableOffset

RecordIdKeyBuilder

RecordIdKeyRange

A range of record ID keys with start and end bounds.

RecordIdKeyRangeArgs

RecordIdKeyRangeBuilder

RecordIdKeyType

A union of all possible record ID key types.

RecordIdKeyTypeUnionTableOffset

RecordKind

Represents the ‘record’ type, optionally constrained to specific tables.

RecordKindArgs

RecordKindBuilder

RegexKind

Represents the ‘regex’ type.

RegexKindArgs

RegexKindBuilder

Set

A unique collection of values.

SetArgs

SetBuilder

SetKind

Represents the ‘set’ type (unique collection of values).

SetKindArgs

SetKindBuilder

StringKind

Represents the ‘string’ type.

StringKindArgs

StringKindBuilder

StringValue

String value.

StringValueArgs

StringValueBuilder

TableKind

Represents the ‘table’ type.

TableKindArgs

TableKindBuilder

Timestamp

Timestamp represented as seconds and nanoseconds since epoch.

TimestampArgs

TimestampBuilder

UInt64Value

64-bit unsigned integer value.

UInt64ValueArgs

UInt64ValueBuilder

Uuid

UUID value.

UuidArgs

UuidBuilder

UuidKind

Represents the ‘uuid’ type.

UuidKindArgs

UuidKindBuilder

Value

A value container that can hold any SurrealDB value type.

ValueArgs

ValueBound

A bound for a range value, which can be inclusive, exclusive, or unbounded.

ValueBoundUnionTableOffset

ValueBuilder

ValueType

A union of all possible value types in SurrealDB.

ValueTypeUnionTableOffset

Enums

AnyKindOffset

ArrayKindOffset

ArrayOffset

BoolKindOffset

BoolValueOffset

BytesKindOffset

BytesOffset

DatetimeKindOffset

DecimalKindOffset

DecimalOffset

DurationKindOffset

DurationOffset

EitherKindOffset

FileKindOffset

FileOffset

Float64ValueOffset

FloatKindOffset

FunctionKindOffset

GeometryCollectionOffset

GeometryKindOffset

GeometryOffset

Int64ValueOffset

IntKindOffset

KeyValueOffset

KindOffset

LineStringOffset

LiteralArrayOffset

LiteralDiscriminatedObjectOffset

LiteralKindOffset

LiteralObjectOffset

MultiLineStringOffset

MultiPointOffset

MultiPolygonOffset

NullKindOffset

NullValueOffset

NumberKindOffset

ObjectFieldOffset

ObjectKindOffset

ObjectOffset

OptionKindOffset

PointOffset

PolygonOffset

RangeKindOffset

RangeOffset

RecordIdKeyOffset

RecordIdKeyRangeOffset

RecordIdOffset

RecordKindOffset

RegexKindOffset

SetKindOffset

SetOffset

StringKindOffset

StringValueOffset

TableKindOffset

TimestampOffset

UInt64ValueOffset

UuidKindOffset

UuidOffset

ValueOffset

Constants

ENUM_MAX_GEOMETRY_KIND_TYPEDeprecated

ENUM_MAX_GEOMETRY_TYPEDeprecated

ENUM_MAX_KIND_TYPEDeprecated

ENUM_MAX_LITERAL_TYPEDeprecated

ENUM_MAX_RECORD_ID_KEY_BOUNDDeprecated

ENUM_MAX_RECORD_ID_KEY_TYPEDeprecated

ENUM_MAX_VALUE_BOUNDDeprecated

ENUM_MAX_VALUE_TYPEDeprecated

ENUM_MIN_GEOMETRY_KIND_TYPEDeprecated

ENUM_MIN_GEOMETRY_TYPEDeprecated

ENUM_MIN_KIND_TYPEDeprecated

ENUM_MIN_LITERAL_TYPEDeprecated

ENUM_MIN_RECORD_ID_KEY_BOUNDDeprecated

ENUM_MIN_RECORD_ID_KEY_TYPEDeprecated

ENUM_MIN_VALUE_BOUNDDeprecated

ENUM_MIN_VALUE_TYPEDeprecated

ENUM_VALUES_GEOMETRY_KIND_TYPEDeprecated

ENUM_VALUES_GEOMETRY_TYPEDeprecated

ENUM_VALUES_KIND_TYPEDeprecated

ENUM_VALUES_LITERAL_TYPEDeprecated

ENUM_VALUES_RECORD_ID_KEY_BOUNDDeprecated

ENUM_VALUES_RECORD_ID_KEY_TYPEDeprecated

ENUM_VALUES_VALUE_BOUNDDeprecated

ENUM_VALUES_VALUE_TYPEDeprecated